requirements.txt
tensorflow-cpu
numpy==1.18.5
Pillow
opencv-python
python-multipart
fastapi
uvicorn
